chore: add intex file to utils + rename CatalogPage
