+++ src/App.tsx
@@ -1,5 +1,6 @@
 import { useState } from 'react';
 import { Outlet } from 'react-router-dom';
+
 import { Footer, Header } from 'components';
 
 import './index.scss';
