+++ src/components/BurgerMenu/BurgerMenu.tsx
@@ -1,8 +1,6 @@
-/* eslint-disable jsx-a11y/no-noninteractive-element-interactions */
-/* eslint-disable jsx-a11y/click-events-have-key-events */
-/* eslint-disable max-len */
 import { FC } from 'react';
 import { NavLink } from 'react-router-dom';
+
 import logo from 'assets/img/logo.svg';
 import { ReactComponent as CloseSvg }
   from 'assets/img/icons/close_icon.svg';
@@ -11,6 +9,7 @@ import { ReactComponent as FavSvg }
 import { ReactComponent as CartSvg }
   from 'assets/img/icons/shopping-bag_icon.svg';
 import classNames from 'classnames';
+
 import styles from './BurgerMenu.module.scss';
 
 type Props = {
