+++ src/components/Card/Card.tsx
@@ -1,11 +1,11 @@
-import React, { useState, useContext } from 'react';
+import { useState, useContext } from 'react';
 import { NavLink } from 'react-router-dom';
 
-import { BASE_URL } from 'utils/fetchProducts';
-import { ButtonType, Phone } from 'types';
-
 import { ProductsContext } from 'context';
-import { Button } from 'components/UI';
+import { Button } from 'components';
+import { BASE_URL } from 'utils';
+
+import { ButtonType, Phone } from 'types';
 
 import styles from './Card.module.scss';
