+++ src/components/Footer/Footer.tsx
@@ -3,7 +3,9 @@ import { NavLink } from 'react-router-dom';
 
 import logo from 'assets/img/logo.svg';
 import { ReactComponent as ArrowSvg } from 'assets/img/icons/chevron-up_icon.svg';
-import { BackToTopArrow } from 'components/BackToTopArrow';
+
+import { BackToTopArrow } from 'components';
+
 import styles from './Footer.module.scss';
 
 export const Footer: React.FC = () => {
