+++ src/components/Header/Header.module.scss
@@ -1,7 +1,7 @@
 .header {
   display: flex;
-  justify-content: space-between;
   align-items: center;
+  justify-content: space-between;
   border-bottom: 1px solid var(--secondaryTextColor);
 
   &__left {
