+++ src/components/PageTitle/PageTitle.tsx
@@ -1,12 +1,10 @@
-import { FC } from 'react';
-
 import styles from './PageTitle.module.scss';
 
 type Props = {
   title: string;
 };
 
-export const PageTitle: FC<Props> = ({ title }) => (
+export const PageTitle: React.FC<Props> = ({ title }) => (
   <div className={styles.pageTitle}>
     <h1
       className={styles.pageTitle__header}
