--- src/components/PaginationNumbering/index.ts
@@ -1,1 +1,0 @@
-export { PaginationNumbering } from './PaginationNumbering';
