+++ src/index.tsx
@@ -1,8 +1,10 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
+
 import { ProductsProvider } from 'context';
-import { Root } from './Root';
-import './index.scss';
+import { Root } from 'Root';
+
+import 'index.scss';
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement,
