+++ src/pages/PhonesPage/PhonesPage.module.scss
