+++ src/pages/RightsPage/RightsPage.tsx
@@ -1,4 +1,4 @@
-import { PageTitle } from 'components/PageTitle';
+import { PageTitle } from 'components';
 // import styles from './RightsPage.module.scss';
 
 export const RightsPage = () => (
