+++ src/pages/index.ts
@@ -1,11 +1,10 @@
 export { HomePage } from './HomePage';
-export { FavouritesPage } from './FavouritesPage';
 export { CartPage } from './CartPage';
-export { CatalogPage } from './CatalogPage';
-export { PhonesPage } from './PhonesPage';
-export { TabletsPage } from './TabletsPage';
-export { AccessoriesPage } from './AccessoriesPage';
-export { ItemCardPage } from './ItemCardPage';
-export { NotFoundPage } from './NotFoundPage';
-export { ContactsPage } from './ContactsPage';
 export { RightsPage } from './RightsPage';
+export { PhonesPage } from './PhonesPage/PhonesPage';
+export { TabletsPage } from './TabletsPage';
+export { NotFoundPage } from './NotFoundPage';
+export { ItemCardPage } from './ItemCardPage';
+export { ContactsPage } from './ContactsPage';
+export { FavouritesPage } from './FavouritesPage';
+export { AccessoriesPage } from './AccessoriesPage';
