+++ src/utils/index.ts
@@ -1,0 +1,4 @@
+export * from './constants';
+export * from './contactsData';
+export * from './fetchProducts';
+export * from './helpers';
